backend/requirements.txt
Django>=5.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
python-dotenv>=1.0
Pillow>=10.0
drf-spectacular==0.27.0
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.0
faker==22.0.0
uuid6>=2025.0.1
mysqlclient>=2.2
whitenoise>=6.5
gunicorn>=22.0
requests>=2.31.0

# Real-time (WebSocket via Django Channels)
daphne>=4.1
channels>=4.0
channels-redis>=4.2

# Task Queue (Celery com broker Redis)
celery>=5.3
redis>=5.0

# Firebase (Push Notifications)
firebase-admin>=6.0
